+++ src/components/result/Lights.tsx
