+++ src/components/result/Model.tsx
